+++ visual-js/visual-snapshots/src/api/visual-snapshots-api.ts
@@ -9,7 +9,7 @@ export interface CreateVisualSnapshotsParams {
   buildId: string;
 }
 
-export class VisualSnapshots {
+export class VisualSnapshotsApi {
   private api: VisualApi;
 
   constructor(api: VisualApi) {
