+++ visual-js/visual-snapshots/src/app/pdf-handler.ts
@@ -1,5 +1,8 @@
-import { CreateVisualSnapshotsParams, VisualSnapshots } from "../api/api.js";
-import { initializeVisualApi, VisualApiParams } from "../api/client.js";
+import {
+  CreateVisualSnapshotsParams,
+  VisualSnapshotsApi,
+} from "../api/visual-snapshots-api.js";
+import { initializeVisualApi, VisualApiParams } from "../api/visual-client.js";
 import { PdfConverter } from "./pdf-converter.js";
 
 export interface PdfCommandParams
@@ -9,7 +12,7 @@ export interface PdfCommandParams
 export class PdfCommandHandler {
   public async handle(pdfFilePath: string, params: PdfCommandParams) {
     const visualApi = initializeVisualApi(params as VisualApiParams);
-    const visualSnapshots = new VisualSnapshots(visualApi);
+    const visualSnapshots = new VisualSnapshotsApi(visualApi);
     const pdfConverter = new PdfConverter();
 
     const pdfPageImages = await pdfConverter.convertPagesToImages(pdfFilePath);
